Removed the duplication in the code that was caused by all the seekers that had their own getTrack() methods and all had the same variables. Now they have the same variables in the same class removing duplication.
The NullSeeker is no longer a NullSeeker and the javadoc now explains why. It has to do with the fact that the NullSeeker is not interested in the track and has no decorated seekers.
