+++ app/models/seeker/CommentIntensitySeeker.java
@@ -13,7 +13,7 @@ import java.util.Collections;
  * intensity of the track.
  * 
  * @since 07-05-2015
- * @version 15-06-2015
+ * @version 16-06-2015
  * 
  * @author stefan boodt
  * @author tomas heinsohn huala
@@ -21,19 +21,9 @@ import java.util.Collections;
 public class CommentIntensitySeeker extends AbstractSeeker {
 
     /**
-     * The id of the track.
-     */
-    private Track track;
-
-    /**
      * The set of comments of the track.
      */
     private CommentList comments;
-    
-    /**
-     * The Seeker that is decorated by this one.
-     */
-    private final Seeker decorate;
     
     /**
      * The filter to get the scores out of.
@@ -69,15 +59,14 @@ public class CommentIntensitySeeker extends AbstractSeeker {
      */
     public CommentIntensitySeeker(final Track track, final Seeker decorate,
             final CommentContentSeeker filter) {
-        setTrack(track);
-        this.decorate = decorate;
+        super(track, decorate);
         setComments(new CommentRetriever(track.getId()).getComments());
         this.filter = filter;
     }
 
     @Override
     public ScoreStorage calculateScores(final int duration) {
-        ScoreStorage storage = decorate.calculateScores(duration);
+        ScoreStorage storage = getDecorate().calculateScores(duration);
         return updateStorage(duration, storage);
     }
     
@@ -90,7 +79,7 @@ public class CommentIntensitySeeker extends AbstractSeeker {
     private ScoreStorage updateStorage(final int duration, final ScoreStorage storage) {
         Collections.sort(comments);
         final int commentsize = comments.size(); // Done here for efficientcy.
-        final int tracklength = track.getDuration();
+        final int tracklength = getTrack().getDuration();
         final int stepsize = Comment.getPeriod();
         for (int time = 0; time < tracklength; time += stepsize) {
             int count = 0;
@@ -138,24 +127,6 @@ public class CommentIntensitySeeker extends AbstractSeeker {
      */
     protected int getWeight(final Comment comment) {
         return 2 + getFilter().contentFilter(comment.getBody());
-    }
-
-    /**
-     * Getter of the track.
-     *
-     * @return The track
-     */
-    public Track getTrack() {
-        return track;
-    }
-
-    /**
-     * Setter of the track.
-     *
-     * @param track The track
-     */
-    public void setTrack(final Track track) {
-        this.track = track;
     }
 
     /**
